Further refactor
Move result set implementations into their own directories.
