--- file_store/result_set.go
@@ -1,7 +1,0 @@
-// A Data store specific way of storing and managing result sets.
-
-// Velociraptor is essentially a query engine, therefore all the
-// results are stored in result sets. This interface abstracts away
-// where and how result sets are stored.
-
-package file_store
