+++ result_sets/simple/simple_test.go
@@ -16,12 +16,12 @@ import (
 	config_proto "www.velocidex.com/golang/velociraptor/config/proto"
 	"www.velocidex.com/golang/velociraptor/file_store"
 	"www.velocidex.com/golang/velociraptor/file_store/api"
-	"www.velocidex.com/golang/velociraptor/file_store/result_sets"
-	"www.velocidex.com/golang/velociraptor/file_store/result_sets/simple"
 	"www.velocidex.com/golang/velociraptor/file_store/test_utils"
 	"www.velocidex.com/golang/velociraptor/json"
 	"www.velocidex.com/golang/velociraptor/paths"
 	"www.velocidex.com/golang/velociraptor/paths/artifacts"
+	"www.velocidex.com/golang/velociraptor/result_sets"
+	"www.velocidex.com/golang/velociraptor/result_sets/simple"
 	"www.velocidex.com/golang/velociraptor/services"
 	"www.velocidex.com/golang/velociraptor/services/inventory"
 	"www.velocidex.com/golang/velociraptor/services/journal"
@@ -52,7 +52,7 @@ type ResultSetTestSuite struct {
 func (self *ResultSetTestSuite) SetupTest() {
 	var err error
 	self.config_obj, err = new(config.Loader).WithFileLoader(
-		"../../../http_comms/test_data/server.config.yaml").
+		"../../http_comms/test_data/server.config.yaml").
 		WithRequiredFrontend().WithWriteback().
 		LoadAndValidate()
 	require.NoError(self.T(), err)
@@ -259,7 +259,7 @@ type ResultSetTestSuiteFileBased struct {
 func (self *ResultSetTestSuiteFileBased) SetupTest() {
 	var err error
 	self.config_obj, err = new(config.Loader).WithFileLoader(
-		"../../../http_comms/test_data/server.config.yaml").
+		"../../http_comms/test_data/server.config.yaml").
 		WithRequiredFrontend().WithWriteback().
 		LoadAndValidate()
 	require.NoError(self.T(), err)
