+++ result_sets/simple/sink.go
