+++ result_sets/timed/docs.go
