+++ result_sets/timed/reader.go
