+++ services/server_artifacts/flusher.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"time"
 
-	"www.velocidex.com/golang/velociraptor/file_store/result_sets"
+	"www.velocidex.com/golang/velociraptor/result_sets"
 )
 
 func ResultSetFlusher(ctx context.Context, rs_writer result_sets.ResultSetWriter) func() {
