+++ vql_plugins/result_sets.go
@@ -1,6 +1,6 @@
 package plugins
 
 import (
-	_ "www.velocidex.com/golang/velociraptor/file_store/result_sets/simple"
-	_ "www.velocidex.com/golang/velociraptor/file_store/result_sets/timed"
+	_ "www.velocidex.com/golang/velociraptor/result_sets/simple"
+	_ "www.velocidex.com/golang/velociraptor/result_sets/timed"
 )
